Split src into multiple files.
