+++ src/store.js
@@ -1,0 +1,13 @@
+import ko from 'knockout';
+
+const stateObservable = ko.observable();
+
+function setState(state) {
+    stateObservable(state);
+}
+
+function getState() {
+    return stateObservable;
+}
+
+export { setState, getState };
